Move disposing of controllers to one class
